requirements.txt
oauthlib==1.0.3
python-twitter==2.2
requests==2.9.1
requests-oauthlib==0.6.0
Wand==0.4.2
wheel==0.24.0
